move Layout and CouplingMap to transpiler, deprecate qiskit.mapper (#2222)
* move Layout and CouplingMap to transpiler, deprecate qiskit.mapper

* move tests

* add missing files
