+++ qiskit/transpiler/passes/mapping/basic_swap.py
@@ -16,7 +16,7 @@ compatible.
 from qiskit.transpiler.basepasses import TransformationPass
 from qiskit.transpiler.exceptions import TranspilerError
 from qiskit.dagcircuit import DAGCircuit
-from qiskit.mapper import Layout
+from qiskit.transpiler import Layout
 from qiskit.extensions.standard import SwapGate
 from .barrier_before_final_measurements import BarrierBeforeFinalMeasurements
