+++ qiskit/transpiler/passes/mapping/stochastic_swap.py
@@ -18,7 +18,7 @@ from qiskit.transpiler.basepasses import TransformationPass
 from qiskit.transpiler.exceptions import TranspilerError
 from qiskit.dagcircuit import DAGCircuit
 from qiskit.extensions.standard import SwapGate
-from qiskit.mapper import Layout
+from qiskit.transpiler import Layout
 from .barrier_before_final_measurements import BarrierBeforeFinalMeasurements
 # pylint: disable=no-name-in-module, import-error
 from .cython.stochastic_swap.utils import nlayout_from_layout
